% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftest.R
\name{diff_test}
\alias{diff_test}
\title{Diff_test for test causal effect among time-varying Exposure.}
\usage{
diff_test(
  data = data,
  expo_name1 = "x1",
  expo_name2 = "x2",
  expo_name3 = "x3",
  continuous = T,
  boot_no = 1000
)
}
\arguments{
\item{data}{an optional data frame containing the variables in the model.}

\item{x1_name}{the name of pre-outcome exposure}

\item{x3_name}{the name of post-outcome exposure}

\item{x2_name}{the name of exposure between pre-outcome exposure
and post-outcome exposure}

\item{boots_no}{the number of bootstrap}
}
\value{
\code{diff_test} the result of Diff and its 95%CI .
}
\description{
Diff_test for test causal effect among time-varying Exposure.
}
\examples{
a <- rnorm(1000,0,1)
b <- a+rnorm(1000,0,1)
c <- b+rnorm(1000,0,1)
hhh <- data.frame(a,b,c)
hhh$x1 <- ifelse(a<(-0.5),0,ifelse(a>0.5,2,1))
hhh$x2 <- ifelse(b<(-0.5),0,ifelse(b>0.5,2,1))
hhh$x3 <- ifelse(c<(-0.5),0,ifelse(c>0.5,2,1))
con_result <- diff_test (data = hhh, expo_name1 =  'x1', expo_name2 =  'x2',
expo_name3 =  'x3', continuous = F, boot_no =1000)


}
