% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncetscat.R
\name{ncets_cat}
\alias{ncets_cat}
\title{Eliminating the Unmeasured Confounders and Estimating Causal Effect
for categorical outcome.}
\usage{
ncets_cat(
  data = data,
  x1_name = "x1",
  x3_name = "x3",
  y_name = "y",
  boots_no = 1000
)
}
\arguments{
\item{data}{an optional data frame containing the variables in the model.}

\item{x1_name}{the name of pre-outcome exposure}

\item{x3_name}{the name of post-outcome exposure}

\item{y_name}{the name of outcome}

\item{boots_no}{the number of bootstrap}
}
\value{
\code{coefficient} the casual effect coefficients
and its 95%CI for categorical variables.
}
\description{
Eliminating the Unmeasured Confounders and Estimating Causal Effect
for categorical outcome.
}
\examples{
u1 <- rbinom(1000,1,0.5)
px1 <- exp(0.8*u1)/(1+exp(0.8*u1))
x1 <- apply(matrix(px1,nrow=1),2,rbinom,n=1,size=1)
px3 <- exp(0.8*u1)/(1+exp(0.8*u1))
x3 <- apply(matrix(px3,nrow=1),2,rbinom,n=1,size=1)
py<- 0.2*x1+0.4*u1
y<- apply(matrix(py,nrow=1),2,rbinom,n=1,size=1)

data <- data.frame(x1,x3,y)
model <- ncets_cat ( data = data,x1_name = 'x1',
x3_name = 'x3', y_name ='y',boots_no = 1000)
model

}
