% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncetscon.R
\name{ncets_con}
\alias{ncets_con}
\title{Eliminating the Unmeasured Confounders and Estimating Causal Effect
for Continuous Outcome.}
\usage{
ncets_con(
  formula,
  data = data,
  sdmethod = c("normal", "bootstrap", "all"),
  x1_name = "x1",
  x3_name = "x3",
  boots_no = NULL
)
}
\arguments{
\item{formula}{an object of class 'formula' :
a symbolic description of the model to be fitted}

\item{data}{an optional data frame containing the variables in the model.}

\item{sdmethod}{Choosing a method for estimating variance}

\item{x1_name}{the name of pre-outcome exposure}

\item{x3_name}{the name of post-outcome exposure}

\item{boots_no}{the number of bootstrap}
}
\value{
\code{formula} the model

\code{bootstrap} the casual effect coefficients
when the variance is calculated by bootstrap

\code{normal} the casual effect coefficients
when the variance is calculated in theory

\code{quant} the confidence interval
calculated by bootstrap
}
\description{
Eliminating the Unmeasured Confounders and Estimating Causal Effect
for Continuous Outcome.
}
\examples{
  u <- rnorm(1000,0,1)
  x1 <- 0.5*u +rnorm(1000,0,1)
  x3 <- 0.5*u +rnorm(1000,0,1)
  y <- 2*x1 + 1*u +rnorm(1000,0,1)
  data <- data.frame(u,x1,x3,y)
  model <- ncets_con (y~x1+x3, data = data, sdmethod ='normal',
  x1_name = 'x1',x3_name = 'x3',boots_no = NULL)
  model

}
